Move add_url_params to c2cgeoportal_commons
