+++ geoportal/c2cgeoportal_geoportal/views/pdfreport.py
@@ -34,8 +34,8 @@ from pyramid.httpexceptions import HTTPBadRequest, HTTPForbidden
 from pyramid.view import view_config
 
 from c2cgeoportal_commons import models
+from c2cgeoportal_commons.lib.url import add_url_params
 from c2cgeoportal_commons.models import main
-from c2cgeoportal_geoportal.lib import add_url_params
 from c2cgeoportal_geoportal.lib.caching import NO_CACHE
 from c2cgeoportal_geoportal.lib.layers import get_private_layers, get_protected_layers
 from c2cgeoportal_geoportal.views.ogcproxy import OGCProxy
