+++ geoportal/c2cgeoportal_geoportal/views/theme.py
@@ -50,10 +50,9 @@ from sqlalchemy.orm import subqueryload
 from sqlalchemy.orm.exc import NoResultFound
 
 from c2cgeoportal_commons import models
-from c2cgeoportal_commons.lib import get_url2
+from c2cgeoportal_commons.lib.url import get_url2, add_url_params
 from c2cgeoportal_commons.models import main
 from c2cgeoportal_geoportal.lib import (
-    add_url_params,
     get_roles_id,
     get_typed,
     get_types_map,
